[BUGFIX] RTE: Mixed up configuration options of the Acronym button
The Acronym button will be enabled even when no acronym record exists.
The Abbreviation tab will be shown even if no acronym record of this
type exists; in such case, the predefined abbreviation fieldset will not be shown.
However, the Acronym tab will not be shown if no acronym record of this
type exists. If the current selection is empty and not within an abbr
tag, the Abbreviation fieldset will not be show. If the current
selection is empty and not within an acronym tag, the Acronym fieldset
will not be show.

Releases: 6.2
Resolves: #64165
Change-Id: I590acd8d2ee10f2e9509ecb15c9b65876a363eda
Reviewed-on: http://review.typo3.org/36725
Tested-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>