[BUGFIX] DatePicker can't handle minDate and maxDate
The Bootstrap DatePicker expects the min and max dates to be formatted
in the format defined in the options. Currently both dates are filled
with an unix timestamp. These can't be converted to a date object (moment
object) and the date picker is completely disabled.

This patch adds the conversion of unix timestamp to the right format
before the Bootstrap DatePicker is initialized.

Releases: master
Resolves: #65068
Change-Id: Ie8143efb537067dd89cfb42aa477f63b8c07f693
Reviewed-on: http://review.typo3.org/36892
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>