[BUGFIX] Prevent Javascript loop for RTE-enabled fields
Problem: If a collapsed IRRE element with a RTE-enabled field is added,
a JavaScript loop is triggered because of missing widths of the select
fields within the RTE toolbar. This is caused by change in the way IRRE
nested inline elements are made visible using class attribute rather than
display property of style attribute. This also applies to new tab menus.
The change also affects the properties of the nested elements whose
mutations must be observe in order to detect when the RTE iframe becomes
visible.
Solution: Adjust the list of attributes to be observed on nested inline
elements. Adjust also detection of nested elements visibility state as
well as the walk to make them hidden with the visibility property of
the style attribute while the RTE framework is being rendered.

There is also a check added for TYPO3.settings.Textarea that is not
present when RTE fields are hidden by DisplayCond for a child element
that gets expanded/opened.

Releases: master
Resolves: #64930
Change-Id: I6f3bea4e9e72c33f2ae773734e9e6d794fe38ee9
Reviewed-on: http://review.typo3.org/36860
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>