[BUGFIX] Install tool fatal with options for cache_classes
Setting an option like cacheDirectory for cache_classes or
cache_core triggers a fatal if calling the install tool.
This happens because the install tool calls the bootstrap
loadConfigurationAndInitialize() with $allowCaching = FALSE
which then calls disableCoreAndClassesCache() and sets the
backend of both caches to NullBackend. The cacheDirectory
option however is not supported by this backend.
Solution is to unset any options configured for these caches
within disableCoreAndClassesCache().

Resolves: #65153
Releases: master, 6.2
Change-Id: I3200d6752794c1769fe4e049c301ee8fa185ecf9
Reviewed-on: http://review.typo3.org/37000
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>