[BUGFIX] Install tool error message if session file not writeable
In a rare case it may happen that the install tool created a session
file in typo3temp/InstallToolSessions and later those files can not
be written or updated due to permission problems.
The install tool then fails silently without error message.
Since write out of session data happens in __destruct() a proper
error message can not be created at this point anymore. Solution
is to "test write" the session file after opening it and throw
an exception if that fails.

Resolves: #65014
Releases: master, 6.2
Change-Id: I9478f0cfb58a047842d48350b7005e64f50f52ff
Reviewed-on: http://review.typo3.org/37028
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>