[BUGFIX] Use current domain if found in domain record
The introduction of domain record cache with #24389 introduced a regression.
Previously if multiple domain names were found for one subtree,
the domain name which matched the current host name was preferred.
After the introduction of the cache, only the first domain record was
considered.

This patch properly checks the domain records against the current domain
in use and selects the most appropriate one.

Releases: master, 6.2
Resolves: #62556
Change-Id: I9cae580af9c9de0c1f6ed02d5ac8ad6a90a0fa2b
Reviewed-on: http://review.typo3.org/37069
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>