[TASK] Replace GeneralUtility::inList() with isset() within loops
Rule of thumb should be to use GeneralUtility::inList() for
single checks of lists with unknown values only. Within loops
it will be faster to explode the list outside the loop, then flip
the array and finally use isset on the array keys within the loop.

Hardcoded lists might be replaced with arrays as well.

See the issue on forge for a description and some benchmarks.

This will not fully resolve the issue since there will be more
patches to come, since we should avoid monster patches to minimize
the risk of regressions.

Releases: master
Resolves: #55626
Change-Id: I5a0d8ea6f71a4f8fd39587c5ec3bfcb557d4c777
Reviewed-on: http://review.typo3.org/37138
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>