[FEATURE] TypoScript condition add GPmerged
If one uses TypoScript condition with GP then the check is with
GeneralUtility::_GP() which will means that if I have GET variables
beginning with an extbase plugin-namespace and POST variables with
the same plugin-namespace, e.g.
GET: tx_demo_demo[action]=detail
POST: tx_demo_demo[name]=Foo

then GeneralUtility::_GP('tx_demo_demo'), as intended, will only
return the array of the POST variables for that namespace.

However, that results in the issue that if I check for the GET variable
the check will fail.

So for such cases there is a new TypoScript condition
GPmerged which uses GeneralUtility::_GPmerged()

Change-Id: I772c1c1d178b10e9293132bc41bfc36fabd53063
Resolves: #65250
Releases: master
Reviewed-on: http://review.typo3.org/37157
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>