[BUGFIX] BE: DatabaseTreeDataProvider maxLevels incorrect
The current behaviour for maxLevels with treeConfig is incorrect.
Currently if one uses maxLevels 0, instead of only seeing the root category,
one actually is presented of nodes up to level 2
(if one would specify the levels in nonSelectableLevels).
With maxLevels 1 it is already up to level 3.
Of course, maxLevels 0 should only display the root level, maxLevels 1 only the first child nodes.
Also increased default levelMaximum to 4, in order to ensure old non configured
trees are still behaving as intended.

Change-Id: Ic1565e39abad01599dd67bbb05b0ce9fa7161af4
Resolves: #65304
Releases: master
Reviewed-on: http://review.typo3.org/37196
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>