[BUGFIX] Properly handle XLF translation file detection
The GeneralUtility::llXmlAutoFileName() method will now also correctly
handle paths to files that are not placed within a known directory
(ext, sysext, etc.) if the call is made with $sameFile = TRUE.
This allows placing language files in arbitrary storage.

This change finally re-enables the usage of locallang files in TypoScript
or f:translate-ViewHelpers which use the XLIFF format and are not placed
within an extension.
This has always been working with files in the XML format.

Resolves: #65273
Releases: master, 6.2
Change-Id: Ibb647cea32a9fc04e9dfdf707a5e8690ad399d5c
Reviewed-on: http://review.typo3.org/37475
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>