[BUGFIX] Allow file:// links in Element Browser
Due to FAL checks it's not possible to link to a file:// resource as
external link. This patch checks for correct "file:" prefix for FAL
relations and allows the use of file:// as external link scheme.

Furthermore a Javascript error is resolved which prevents the closing
of the Element Browser in Chrome.

Releases: master, 6.2
Resolves: #65358
Change-Id: I706fd975654bf82a0830d9901f08ae9de7cc49bf
Reviewed-on: http://review.typo3.org/37329
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>