[BUGFIX] Invalid back path of ImageViewHelper in TYPO3 backend
If rendering an image in the TYPO3 backend, the regular case is
to prepend "../" to an URI with the assumption that the original
HTTP request was handled by typo3/backend.php.

For the page module for instance, the request points to
typo3/sysext/cms/layout/db_layout.php, thus the back path way
does not work anymore.

To overcome that behavior the absolute path shall be used
instead by determining TYPO3_SITE_PATH. For the case that
some custom processing relies on "../" this change might
be breaking.

Resolves: #65433
Releases: master, 6.2
Change-Id: I969f6a3bdd000b7efb7607e533d1117c8d363959
Reviewed-on: http://review.typo3.org/37408
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>