[BUGFIX] Resolve class name from aliases managed by composer
The old class loader has API to return the original class name
when given an alias. This API was not respected when implementing
composer class loading with #64361 with the result that it was
broken for all classes and aliases managed by the composer class loader.

To address this issue, the composer class loader is now injected into
the old class alias map, so that it can resolve class names from aliases
managed by it.

Related: #64361
Resolves: #65667
Releases: 6.2, master
Change-Id: If0b5098d72147fe7d76e97ee61c81d87cad72b4b
Reviewed-on: http://review.typo3.org/37825
Reviewed-by: Sascha Nowak <sascha.nowak@netlogix.de>
Tested-by: Sascha Nowak <sascha.nowak@netlogix.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>