[BUGFIX] Hooks/Signals cannot be used during extension install
During the installation of an extension in the extension manager
hooks of the extension to be installed cannot be used currently.
This behavior gets handy if e.g. data to be imported already
should be processed by a custom handler.

By restructuring the process, especially flushing caches and
reloading them, this can be achieved.

Resolves: #65848
Releases: master, 6.2
Change-Id: I89f1fed988377d0a38c6a5908e1a297342754c36
Reviewed-on: http://review.typo3.org/37955
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>