[!!!][TASK] Do not add JSON to header if type equals "json"
If you choose "json" as contentType in AJAX-Calls the encoded
json-string will be pushed to the HTTP-Header and will be
outputted on the website. But HTTP-Header is not the right
place for huge json-strings.
So this patch lets "json" works like "jsonbody".

Back in the days, JSON data in the X-JSON header
was for prototypeJS usage, which is now not used
anymore for AJAX requests within the core.

Resolves: #50509
Releases: master
Change-Id: I782cdf2817e30f2e5f016fa92edb06f1e8cbd5b2
Reviewed-on: http://review.typo3.org/22655
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>