[BUGFIX] Enable page tree for editors
Currently for editors the request for page tree actions fails with error
"Access Error: You don't have access to this module.". This is because
there is an access check to web module, which can't be enabled for users
anymore. As the page tree checks access by user and group definitions,
the scripts don't need any check themselves.

Resolves: #65972
Releases: master
Change-Id: I57ef845508eaa64bd2156f793a13ace7a39b2cbd
Reviewed-on: http://review.typo3.org/38110
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>