[BUGFIX] Indexed_search extbase plugin support for mysql fulltext search
Add three getters for IndexSearchRepository and SearchFormController
to have a common api for accessing data needed by FulltextIndexHook.

Unify the processing of $page_where as in one class it
was expected that it returns statement with 'AND' and in other without.

Unify the formatting of select statements in both classes to make
differences easier to spot (fix missing $wordSel statement
in the extbase plugin in the process).

Resolves: #65989
Resolves: #52277
Releases: master, 6.2
Change-Id: I9fcb86d44208e8fe6d205667f344022b0ad62505
Reviewed-on: http://review.typo3.org/38744
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>