[TASK] Remove unused icons - part 2
All files in EXT:t3skin/icons/gfx/ and typo3/gfx/ which are not
in use anymore (by LESS/CSS) are removed. All files which
are still in use, but are overridden by t3skin through
"skinImgAutoCfg" are moved to typo3/gfx/ to replace the
old (pre 4.0) default files.

Releases: master
Resolves: #66000
Change-Id: I0591d951cd7295041a05a963efa6321403c66642
Reviewed-on: http://review.typo3.org/38136
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>