[BUGFIX] Performance issue with caching empty text menus
If a text menu level is defined in TypoScript but does not
return any pages, the result value is NULL instead of the
expected empty array. Since the check whether to reprocess
a menu expects an array, the generation process is triggered
again, invalidates the cache entry and adds the very same NULL
value to the cache.

The solution is to define a default empty array value for text
menu levels.

Resolves: #66023
Releases: 6.2
Change-Id: I66abb6fe7c1a91544b02193573ec601fb3240d98
Reviewed-on: http://review.typo3.org/38159
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>