[TASK] Make Backend Module URLs absolute
In the approach to remove all occurences of dealing with
relative URL referencing in the TYPO3 Backend, the main mod.php
calls are now always prefixed with "/{possiblesubfolder/}typo3/" so
no backPath parameter and calculations are needed anymore.

The patch thus removes all locations of the JavaScript
variable "top.TS.PATH_typo3" containing the backpath and
makes the "relative" path to BackendUtility::getModuleUrl()
always a absolute prefix relative to the PATH_typo3 structure.

Resolves: #66052
Releases: master
Change-Id: Ie8192f227b8d1715be90efff8746482adf252ac7
Reviewed-on: http://review.typo3.org/38246
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>