[BUGFIX] Last-Modified response header broken by design
The Last-Modifed header needs to change whenever there is a change
related to that page. Not only the page properties or records located
on that page, but every change which should be reflected on that page,
like menus, TypoScript or layout. Currently this is not happening and
has huge impact to make it work.

This patch removes the Last-Modified response header. The ETag response
header has kind of the same function, so no harm is done by deleting
Last-Modified.

Resolves: #63927
Releases: master, 6.2
Change-Id: Ic5d14225c2f90cc63e1c4413d5426565ea5e1e41
Reviewed-on: http://review.typo3.org/38269
Reviewed-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Tested-by: Andreas Fernandez <andreas.fernandez@aspedia.de>