[BUGFIX] Add imageMagickIdentify fallback to ImageInfo::getImageSizes()
ImageInfo::getImageSizes() uses getimagesize() to determine the image
dimensions. But for pdf and svg files this fails.

This patch add a fallback to GraphicalFunctions::imageMagickIdentify()
when getimagesize() doesn't give any results.

This patch also changes the Indexer::extractRequiredMetaData() as is
had the same issue. It now used the ImageInfo object.

Resolves: #66268
Releases: master, 6.2
Change-Id: I51d8ee5106b07667139f49a7acdfaabf99501a5b
Reviewed-on: http://review.typo3.org/38476
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>