[BUGFIX] Make global TSFE and TT a reference to the class members
The global variables $GLOBALS['TSFE'] and ['TT'] are modified in
some places throughout FE rendering.
Since those public objects should only be mirrors of the protected
members of the RequestHandler, we need to make the global variables
a reference to the protected members.

Be aware that this is a dirty workaround and actually bypasses
the protected access of the members by making them changable
via the global variables.

Change-Id: I542f382ee42bdbacc6bad7d74dfb69dfe6fcca88
Resolves: #66278
Relates: #65914
Releases: master
Reviewed-on: http://review.typo3.org/38489
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Tested-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>