[BUGFIX] Make sure excludedTablesAndFields are empty in DataHandler
Simulating the admin function when saving their own account failed when
TCA is modified. The backend user is now simulated as admin, instead of
the DataHandler class only.

Now it validates all possible checks when updating the user records as
admin.

Note: This will keep the same functionality and therefor has no impact
on the Core itself.

Releases: master, 6.2
Resolves: #66412
Change-Id: Ib17ac99d9e7a4aaf1e807b41b30f7310ad8da655
Reviewed-on: http://review.typo3.org/39049
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>