[BUGFIX] Calculate correct cropping for large images
When the visible image in the cropping wizard is a scaled
version of the original image we need to take the scaled ratio into
account when setting the crop info.

Scaling of the original images occurs when it is bigger then
1000px x 700px.

This patch makes sure the crop info is correctly calculated.

Resolves: #66440
Releases: master
Change-Id: I7852b19a47f462120c6ea4b72ec388254a210d3e
Reviewed-on: http://review.typo3.org/38717
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Tested-by: Andreas Fernandez <andreas.fernandez@aspedia.de>