[BUGFIX] FormEngine exception and warning creating new record
Some extensions like ext:news use ';;;;2-2-2' in TCA showItem which is
actually an invalid syntax. FormEngine detects this and throws an
exception. To mitigate the issue, the exception is commented out for
now and a @todo is added on how this case could be handled later.

Another warning in flex form container / section handling is fixed
with the patch by sanitizing the given row data to an array, which
is not always the case, for example if adding a new record.

Resolves: #66628
Related: #63846
Releases: master
Change-Id: Ib6dcba0c5e2c1dad8c842e99994ab97e456deb41
Reviewed-on: http://review.typo3.org/39017
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>