[TASK] Travis: Mark functionals as allowed_failure
With mysql 5.6 we still see various mysql issues that come from
travis infrastructure. We're currently unable to fix this. To have
a green build for now, functional tests are separated and marked
as allowed_failure until stuff is sorted out on travis side.
Patch is for master only at the moment, 6.2 may get an own change
if problems persist.

Releases: master
Change-Id: I9513c8894d7737c09a47588dc9efddd4e9e140e3
Reviewed-on: http://review.typo3.org/39073
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>