[BUGFIX] Generate thumbnails for files without width/height
Thumbnails should not upscale images, therefore we check for the
original size against thumbnail size. If original size is 0 (as for
pdf-Files) this leads to not generating thumbnails for those files. Now
checking original width/height against 0 to prevent this and generate
thumbs for those files again.

Resolves: #66679
Related: #66270
Releases: master, 6.2
Change-Id: I5823142a6be19e6068fea358027dc808f5fd3185
Reviewed-on: http://review.typo3.org/39077
Tested-by: Markus Klösges <mkloesges@gmx.de>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>