[BUGFIX] Reload classAliasMap after extension installation
If an extension is installed at least the system cache is cleared
which throws away all defined class aliases as well. Those are only
loaded by bootstrap at the very beginning of an request and need to
be reloaded by the install process manually. This patch adds the
necessary call to ClassLoader.

Releases: master, 6.2
Resolves: #66696
Related: #65141
Change-Id: I4ace91789c43855acba91ac356b2263171cc921a
Reviewed-on: http://review.typo3.org/39112
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>