[BUGFIX] Prevent sorting incomplete loaded packages
If an extension is defined in PackageStates.php but missing in file
system the package gets unloaded and the PackageStates.php is rewritten.
If a core extension is missing (e.g. extra_page_cm_options) the file
is rewritten with incomplete loaded package information and gets wrongly
sorted.

Releases: master, 6.2
Resolves: #66699
Change-Id: I18e4bb83cf1676537a45ea4b07b9228335e34ec7
Reviewed-on: http://review.typo3.org/39147
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>