[BUGFIX] Prevent wrong information after extension upload
If you upload an extension it is extracted and installed in separate
steps. If the extraction was successful a flash message is rendered.
If the installation fails the whole extension is removed again but the
flash message is still visible. This patch separates flash message output
for upload and installation process.

Releases: master, 6.2
Resolves: #66729
Change-Id: I6d84d8ecebd55a989a34c1024300916e33af66a7
Reviewed-on: http://review.typo3.org/39181
Reviewed-by: Jan Helke <typo3@helke.de>
Tested-by: Jan Helke <typo3@helke.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>