[BUGFIX] Activate runtimeActivatedPackages without cache clearing
Currently you have to clear all caches manually to use runtime activated
packages because they are not included in the building process of the
cache identifier. This patch defines the TYPO3_CONF_VARS setting per
default and includes it in the cache identifier. Furthermore runtime
activated packages are excluded from being disabled in EM.

Release: master, 6.2
Resolves: #66768
Resolves: #58874
Change-Id: Idbda3c4f1f3c827c2b26c6f7759a9121cc260068
Reviewed-on: http://review.typo3.org/39378
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>