[TASK] FormEngine: Simplify placeholder and NULL handling
FormEngine / DataHandler have logic to distinguish between NULL and
empty strings as values written to the database. This is especially
used for FAL in inline relation together with the placeholder logic.
The patch moves the code around a bit, disentangles the logic and
adds comments to explain the details. The $globalOptions value
parameter "prependFormFieldNamesActive" becomes obsolete along the
way and is dropped.

Change-Id: I854a5733f9b597a28751ed31217858b223441247
Resolves: #66856
Releases: master
Reviewed-on: http://review.typo3.org/39365
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>