[BUGFIX] Temporary filename collision in imageMagickConvert()
Add more entropy to the temporary filename used by imageMagickConvert()
to ensure different processing configurations resulting in the same
output dimensions get different temporary names.

In combination with LocalDriver moving the temporary file to a different
location and file_exists_typo3temp_file() having a 30s window where it
doesn't detect the moved file (when the user enabled
$TYPO3_CONF_VARS['GFX']['enable_typo3temp_db_tracking']) this
leads to entries in the sys_processed_file table with a width and
height of 0, thus referencing the original file and resulting in
unscaled images showing in the frontend.

Resolves: #65842
Resolves: #63519
Resolves: #60215
Related: #65378
Releases: 6.2
Change-Id: I42748d0899bf1e24f9f92f2e55802d64309c5704
Reviewed-on: http://review.typo3.org/39373
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>