[TASK] Sane index length for "uid_foreign_tablefield"
The index lengths for the fields "tablenames" and "fieldname"
of table "sys_category_record_mm" are way too big.
It suffices to have only few significant characters in the index.

We only take the first 40 characters of the tablename and the first
three characters of the fieldname as those should already be
distinctive enough to have a high quality index.
Assuming most table names are below 40 characters anyway and usually
only a few fields per table will be a category.

Resolves: #66773
Releases: master
Change-Id: Ib1f077c93425992c557fbc4457699d867521a3ba
Reviewed-on: http://review.typo3.org/39443
Reviewed-by: Stephan Großberndt <stephan@grossberndt.de>
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Stefan Froemken <froemken@gmail.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>